CAMEL-6476: Introducing StreamCachingStrategy SPI to make it easier to configure and allow 3rd party to plugin custom strategies.

+++ camel-core/src/main/java/org/apache/camel/api/management/mbean/ManagedStreamCachingStrategyMBean.java
@@ -40,7 +40,7 @@ public interface ManagedStreamCachingStrategyMBean {
     void setSpoolUsedHeapMemoryThreshold(int percentage);
 
     @ManagedAttribute(description = "Percentage (1-99) of used heap memory threshold to activate spooling to disk")
-    long getSpoolUsedHeapMemoryThreshold();
+    int getSpoolUsedHeapMemoryThreshold();
 
     @ManagedAttribute(description = "Buffer size in bytes to use when coping between buffers")
     void setBufferSize(int bufferSize);
